resolving/binding: 11.1 - 11.4
one translation pain was because the idea of storing a dict with object keys
doesn't work in Julia. Structs are by default immutable, and can be compared
based on equality of their contents. Immutable structs have unique pointers
and so they can do the desired kind of comparison.

This led to subtle bugs where a for loop (all variables used in same line,
so they had the same Expression in the dict) didn't behave the same as a while
loop.

I was able to get at the issue by realizing that

```lox
for (var i=0; i<=5; i++) print i;
```

didn't work, but

```lox
for (
	var i=0;
	i<=5;
	i++
) print i;
```

did work!

I also printed out the `locals` which showed me the difference in how those
two were being resolved previously.